+++ src/Copse.Core.Async/AsyncTreeWalker.cs
@@ -3,7 +3,7 @@ using System;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 
-namespace Copse
+namespace Copse.Core
 {
   /// <summary>
   /// A movable stance on a tree: one node together with the ability to navigate away from it --