+++ src/Copse.Linq.Async/TreenumerableExtensions/TreeWalker/AsyncTreeWalker.Subtree.cs
@@ -1,3 +1,4 @@
+using Copse.Core;
 using Copse;
 using Copse.Linq.Treenumerables;