+++ src/Copse.Linq.Async/Treenumerables/Buffers/IAsyncTreenumerableBuffer.cs
@@ -1,3 +1,4 @@
+using Copse.Core;
 using Copse;
 
 namespace Copse.Linq.Treenumerables