+++ src/Copse.Linq.Tests/TreeWalkerLawTests.cs
@@ -1,3 +1,4 @@
+using Copse.Core;
 using Copse.Topologies;
 using Copse.Linq.Treenumerables;
 using Copse;