+++ src/Copse.SimpleSerializer/TreeSerializer.StreamAsync.cs
@@ -1,4 +1,3 @@
-using Copse;
 using Copse.Treenumerables;
 using Copse.Core;
 using System;